packages/shared-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shared-py"
version = "0.1.0"
description = "Librería compartida entre servicios Python del AI Venture Studio"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "nats-py>=2.6",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
